Implemented the HarmonizeTypePass
Before running this pass, when a MMSafePtr is dereferenced, a malformed
load would be generated. For example,

   %p_Obj_Ptr = load %struct.Node*, { %struct.Node*, i64 }* %p

Here we have a type mismatch, which may causes a later pass to fail.
For example, the EarlyCSE pass calls doRAUW() which finds the type
mismatch  between struct.Node* and {struct.Node*, i64}.

This pass will replace this kind of load with a GEP and a new load.
For the example above, it would be replace by

  %Struct_Ptr = getelementptr { %struct.Node*, i64 }, { %struct.Node*, i64 }* %p, i32 0, i32 0
  %loadStructPtr = load %struct.Node*, %struct.Node** %Struct_Ptr