Added code to generate the StructType for a _MM_array_ptr.
A _MM_array_ptr has three fields: the raw pointer to the current
location of an array, an ID for the array, and a pointer to the
start location of the array. The second pointer is used to quickly
locate where the ID of the array is for dynamic ID checking:
currently we put the ID of an array right before the first element
of it.

Also revised the constructor of PointerType a bit. Removed the
default argument for isMMPtr from the constructor and initialize
the default value in the class field declaration.