HarmonizeTypePass fixed ill-formed Extract/InsertValueInst for *p++/--.
When an MMSafePtr is dereferenced or an MMArrayPtr has a
pointer arithmetic, LLVM will generate a clang::CodeGen::Address
to wrap the pointer, and in our current implemenation we let the
compiler mutate the type of the MMSafePtr to the type of the
raw C pointer in the MMSafePtr, i.e., from an llvm::StructTytpe to
an llvm::PointerType. To be more specific, every time an
clang::CodeGen::Address is created for an MMSafePtr, the type of its
llvm::Value *Pointer field will be mutated.
This would cause the generation of ill-formed load and
store instructions. The HarmonizeTypePass handles the case of
pointer dereference.

However, an ill-formed load will also contaminate the ExtractValue
and InsertValue instructions created for MMArrayPtr ++/-- operators.
For example, for *p++/--, the following code would be created:

  %1 = load i32*, { i32*, i64, i64* }* %p, align 32
  %_innerPtr = extractvalue i32* %1, 0
  %incdec.ptr = getelementptr inbounds i32, i32* %_innerPtr, i32 -1
  %2 = insertvalue i32* %1, i32* %incdec.ptr, 0

To fix this, we let the compiler load the whole MMArrayPtr and replace
the polluted one used in ExtractValue and InsertValue.
For the example above, the fixed code looks like:

  %MMArrayPtr = load { i32*, i64, i64* }, { i32*, i64, i64* }* %p
  %_innerPtr = extractvalue { i32*, i64, i64* } %MMArrayPtr, 0
  %incdec.ptr = getelementptr inbounds i32, i32* %_innerPtr, i32 -1
  %1 = insertvalue { i32*, i64, i64* } %MMArrayPtr, i32* %incdec.ptr, 0

Note that although sematically "*p++/--" is equal to "*p; p++",
LLVM generates different IR code for the two situations.
For the latter, the ++/-- operator does not use the type-crippled
MMArrayPtr but uses a complete MMArrayPtr loaded earlier.

TO-DO: handle *++/--p.