HarmonizeTypePass fixed ill-forms instructions from "*++/--p".
"*++/--p" for MMArrayPtr will create ill-formed store instructions.
Because the increment/decrement operator generates an Address for
an MMArrayPtr with mutated type, the following store instruction
(store the updated pointer value) would be broken.
To be more specific, code like the following snnipet would be created:

  %2 = insertvalue { i32*, i64, i64* } %1, i32* %incdec.ptr, 0
  store i32* %2, { i32*, i64, i64* }* %p, align 32

To fix it, first restore the type of the MMSafePtr; this will
automatically fix the store. Next extract the inner raw pointer
from the MMSafePtr and replace all the uses of the ill-typed MMArrayPtr
in load instructions with the extracted one.
For the example, the fixed code would be like:

  %2 = insertvalue { i32*, i64, i64* } %1, i32* %incdec.ptr, 0
  %_innerPtr1 = extractvalue { i32*, i64, i64* } %2, 0
  store { i32*, i64, i64* } %2, { i32*, i64, i64* }* %p, align 32

And the corresponding load that uses the MMArrayPtr would be converted
from "%3 = load i32, i32* %2, align 4"
to   "%3 = load i32, i32* %_innerPtr1, align 4".

Also updated a few variables' names.