Refactored llvm to use "key" and "lock" instead of "ID" for MMSafePtr.
Refactored llvm to use "key" and "lock" rather than "ID" for
temporal memory safety related code. key-lock is semantically clearer
than "ID"; and it is also cleaner: we do not need to differentiate
between the ID of an MMSafePtr and the ID of the memory object.

No functionality changed.

The corresponding change at the clang side is
https://github.com/jzhou76/checkedc-clang/commit/e1e0583d9ec6a57ad97ee832016d693fc262b134