Revised the HarmonizeTypePass.
The old implementation created a new load to load the whole MMArrayPtr
when an ill-formed load is encountered. However, the compiler only need
to generate this load when the ill-formed load is used in instructions
that expect a whole MMArrayPtr, such as ExtractValueInst and
InsertValueInst. This commit delays the construction of this load
to the point where ill-formed ExtractValueInst/InsertValueInst
is found.

Also made minor changes to variable names and comments.