Fixed type mismatch caused by direct dereference of pointer arithmetic.
Directly dereferencing the result of an mm_array_ptr arithmetic would
cause a subtle type mismatch because the result of a pointer arithmetic
is a new mm_array_ptr constructed by InsertValueInst, and the direct
dereference would create an clang::Address that mutates the type of the
new mm_array_ptr. For example, for "*(p + 2)" where p is a
mm_array_ptr<int>, the following IR would be generated (-O0):

    // constructing the new mm_array_ptr
    ...
    %8 = insertvalue { i32*, i64, i64* } %7, i64* %5, 2
    ...
    %9 = load i32, i32* %8, align 4

Here the load is well-formed but we still have a latent type mismatch.
This commit fixed this issue by restoring the type of the
InsertValueInst and replacing the load with a new one.
The fixed instruction sequence would be like:

    %8 = insertvalue { i32*, i64, i64* } %7, i64* %5, 2
    ...
    %_innerPtr2 = extractvalue { i32*, i64, i64* } %8, 0
    %FixedLoad = load i32, i32* %_innerPtr2

Also added split lines to the long comment before the start of the
pass to split the descriptions for different functionalities.