Added the skeleton code for the key check optimization pass.
This commits adds the skeleton code of three passes

- CheckedCKeyCheckOpt
- CheckedCSplitBB
- CheckedCFreeFinder

The second and third are helper passes for CheckedCKeyCheckOpt.
CheckedCSplitBB splits basic blocks by function calls to make the
later analysis easier; and CheckedCFreeFinder finds all functions that
may directly or indirectly free heap objects because during the key
check optimization we need know if a function call may kill the
liveness of the check for an mmsafe pointer.

Currently we plan to implement them as ModulePass and run them
at the beginning of the module pass pipeline. This may change.