Implmeneted the CheckedCFreeFinderPass.
Implemented the pass that finds all functions in a module that may
directly or indirectly free heap objects. It is very conservative
in that if a function has indirect function call(s) or calls a function
defined outside the current compile module, it assumes the function
may free heap. Also if a function may call, directly or indirectly,
another function that meets a condition mentioned in the previous
sentence, it is regarded as a function that may free.

See the code for more details.