Changed the CheckedCFindFreePass to find calls that may cause free.
Modified the CheckedCFreeFinder pass to find function calls that
may directly or indirectly free heap objects. This makes the
CheckedCSplitBB pass much easier to implement. It only needs to
split basic blocks according to the call sites collected from
CheckedCFindFreePass; it saves the efforts of iterating over all
basic blocks to find the call sites of interest.

The comment for the FindMayFreeCalls() function describes the
algorithm of finding the target call sites.