Implemented the CheckedCSplitBB pass.
Implmented the CheckedCSplitBB pass. The output from this pass is
that every basic block either has no function calls that may free or
has at most one function call that may free.

The implementation is pretty straightforward because it uses the
set of may-free functions calls collected from the CheckedCFreeFinder
pass. It simply splits a basic block at the site of a may-free call and
the instruction after the call.