[WIP] Added a function to insert key checks for MMSafePtr argument(s).
Added a function that inserts key check calls for all the MMSafe pointer
argument(s) of a function call. See the comments of the
addKeyCheckForCalls() function for the reasons.

Also moved the key check optimization pass from the beginning of the
pass pileline to right after the mem2reg pass; this pass helps reduce
the number of alloca instructions and conseuquently reduce aliased
pointers. For example, for a local pointer p, before this pass there
might be multiple aliased pointers to p and each of them is created
by an alloca. Without alias analysis the compiler doesn't know
they are aliases and thus may cause missing removing redundant
key checks for p. The mem2reg pass reduces unneccessary allocas.

TODO: update the main body of the OPT() function to reflect the change
made in this commit. Specifically, it should assume that all MMSafePtr
argument(s) have been checked and thus can safely elide the checks for
their first dereferences in the function.