[Key Check Opt] Fixed two bugs.
Fixed two bugs in the key check opt pass.

First, we omitted one case for the source of a pointer argument.
A pointer argument may come from:
  - LoadInst
  - ExtractValueInst (omitted, needs handling)
  - a NULL variable (can be omitted safely)
  - an argument of the caller function (can be omitted safely)
This commit added code to handle the second case. Also when iterating
the arguments of a function call, it added the Value::strpPointerCast()
call for them; it would otherwise miss mmsafe pointer arguments if it
only checks if it's a load or extractvalue instruction.

Second, before inserting a call to a key check function we need check
if the target mmsafe pointer is NULL; there would be a sefault if
a key check function works on a NULL mmsafe pointer.