[_multiple] Fixed a sutble alignment bug in _multiple object generation.
There is an alignment issue in the _multiple object generation. For
MMSafe pointers, we need to guarantee that they are aligned by 16 bytes.
When an mmsafe pointer is declared in a struct in the source code,
Clang guarantees that it is either 16 bytes aligned (mmptr) or
32 bytes aligned (mmarrayptr) because we updated related src code
(include/clang/Basic/TargetInfo.h) for it.  However, when we manually
build a struct that contains a 64-bit integer and an mmsafe pointer,
no padding would be inserted between the integer and the pointer
to guarantee the alignment of the mmsafe ptr field, and it may cause
segfault when a movaps instruction is used to load/store the mmsafe ptr
field. I did not find an API to tweak the alignment of individual
struct fields. Our solution is to insert another 64-bit integer in the
new struct and set the alignment of the whole struct to be 16 bytes.
This solution is not elegant in my opinion, although the generated
binary code should be the same even if there is an API function that
pads the struct for alignment.

Related commits of width and alignment setting for mmsafe pointers:
https://github.com/jzhou76/checkedc-clang/commit/23cea79f3c757c1d59376c4d7f704f9121e4deed
https://github.com/jzhou76/checkedc-clang/commit/a97f89935e549f9a96c60a7e4ac6a02530cae523