[Opt] Fixed a bug in the key check optimization pass.
The "changed" flag used in the while loop to minitor changed sets
of checked pointers should always be updated with an "or" operator
with the old "changed". The old code might incorrectly set a changed
flag to unchanged and thus cause missing certain propagations.
See the code for more details.